Изменения в авторизации и мелкие баги

+++ AccountServer/TwoButtonsAccountDatabase/DTO/UserContactsDTO.cs
@@ -3,6 +3,6 @@
   public class UserContactsDto
   {
     public int VkId { get; set; }
-    public int FacebookId { get; set; }
+    public long FacebookId { get; set; }
   }
 }
